Create launch countdown widget
- Create widget
- Create a draft count in home
- Eliminate launch widget for spacex_home
